+++ SalsaVerde/stock/tests/test_containers.py
@@ -1,16 +1,16 @@
 from datetime import datetime
 
 from django.conf import settings
-from django.test import TestCase
 from django.urls import reverse
 
+from SalsaVerde.common.tests import SVTestCase
 from SalsaVerde.stock.factories.raw_materials import ContainerFactory, ContainerTypeFactory
 from SalsaVerde.stock.factories.supplier import SupplierFactory
 from SalsaVerde.stock.models import Container, ContainerType, Document
 from SalsaVerde.stock.tests.test_common import AuthenticatedClient, refresh
 
 
-class ContainerTypeTestCase(TestCase):
+class ContainerTypeTestCase(SVTestCase):
     def setUp(self):
         self.client = AuthenticatedClient()
         self.user = self.client.user
@@ -63,7 +63,7 @@ class ContainerTypeTestCase(TestCase):
         self.assertContains(r, reverse('container-types-details', args=[ct.pk]))
 
 
-class ContainerTestCase(TestCase):
+class ContainerTestCase(SVTestCase):
     def setUp(self):
         self.client = AuthenticatedClient()
         self.user = self.client.user
@@ -80,41 +80,51 @@ class ContainerTestCase(TestCase):
         data = {
             'intake_date': datetime(2018, 2, 2).strftime(settings.DT_FORMAT),
             'intake_user': self.user.pk,
+            'intake_notes': 'Foobar',
             'container_type': self.container_type.pk,
             'quantity': 10,
             'batch_code': '123abc',
             'supplier': self.supplier.pk,
-            'condition': 'Good',
+            'intake_quality_check': True,
         }
         r = self.client.post(self.add_url, data=data)
         self.assertRedirects(r, reverse('containers'))
-        doc = Document.objects.get()
-        assert doc.type == Document.FORM_SUP02
-        assert doc.author == self.user
+
         container = Container.objects.get()
         assert container.container_type == self.container_type
         assert container.batch_code == '123abc'
-        assert container.condition == 'Good'
         assert container.quantity == 10
         assert container.supplier == self.supplier
+        assert container.intake_notes == 'Foobar'
+        assert container.intake_quality_check
 
         r = self.client.get(reverse('containers-details', args=[container.pk]))
-        self.assertContains(r, f'{reverse("documents-details", args=[doc.pk])}">SUP02')
+        self.assertContains(r, 'Foobar')
 
     def test_edit_container(self):
-        container = ContainerFactory(container_type=self.container_type, batch_code='foo123')
+        container = ContainerFactory(container_type=self.container_type, batch_code='foo123', intake_user=self.user)
         r = self.client.get(reverse('containers-edit', args=[container.pk]))
         self.assertContains(r, 'foo123')
         data = {
-            'container_type': self.container_type.id,
-            'supplier': self.supplier.id,
+            'intake_date': datetime(2018, 2, 2).strftime(settings.DT_FORMAT),
+            'intake_user': self.user.pk,
+            'intake_notes': 'Foobar',
+            'container_type': self.container_type.pk,
+            'quantity': 10,
             'batch_code': '123abc',
-            'quantity': 5,
-            'condition': 'Good',
+            'supplier': self.supplier.pk,
+            'intake_quality_check': True,
         }
         r = self.client.post(reverse('containers-edit', args=[container.pk]), data=data)
         self.assertRedirects(r, reverse('containers-details', args=[container.pk]))
-        assert refresh(container).batch_code == '123abc'
+
+        container = Container.objects.get()
+        assert container.container_type == self.container_type
+        assert container.batch_code == '123abc'
+        assert container.quantity == 10
+        assert container.supplier == self.supplier
+        assert container.intake_notes == 'Foobar'
+        assert container.intake_quality_check
 
     def test_delete_container(self):
         con = ContainerFactory(container_type=self.container_type)
