+++ SalsaVerde/stock/views/container_types/__init__.py
@@ -1,0 +1,3 @@
+from SalsaVerde.stock.views.container_types import details
+from SalsaVerde.stock.views.container_types import form_views
+from SalsaVerde.stock.views.container_types import list
