+++ SalsaVerde/stock/views/containers/__init__.py
@@ -1,0 +1,3 @@
+from SalsaVerde.stock.views.containers import details
+from SalsaVerde.stock.views.containers import form_views
+from SalsaVerde.stock.views.containers import list
