+++ SalsaVerde/stock/views/ingredient_types/__init__.py
@@ -1,0 +1,3 @@
+from SalsaVerde.stock.views.ingredient_types import details
+from SalsaVerde.stock.views.ingredient_types import form_views
+from SalsaVerde.stock.views.ingredient_types import list
