+++ SalsaVerde/stock/views/ingredients/__init__.py
@@ -1,0 +1,3 @@
+from SalsaVerde.stock.views.ingredients import details
+from SalsaVerde.stock.views.ingredients import form_views
+from SalsaVerde.stock.views.ingredients import list
