+++ SalsaVerde/stock/views/products/__init__.py
@@ -1,0 +1,3 @@
+from SalsaVerde.stock.views.products import details
+from SalsaVerde.stock.views.products import form_views
+from SalsaVerde.stock.views.products import list
