+++ static/scss/variables.scss
@@ -18,8 +18,6 @@ $link-color:                #3062bb;
 $card-bg:                   white;
 $card-title-spacer-y:       0;
 
-//$link-color:                $main-colour;
-//$color
 
 @import 'DL/bootstrap-scss/functions';
 @import 'DL/bootstrap-scss/variables';
