+++ templates/macros.jinja
@@ -166,3 +166,11 @@
     </nav>
   {% endif %}
 {% endmacro %}
+
+{% macro spinner() %}
+  <div class="d-flex justify-content-center">
+    <div class="spinner-border text-primary" role="status">
+      <span class="visually-hidden">Loading...</span>
+    </div>
+  </div>
+{% endmacro %}
